[nrf noup] boards: add nrf54l15/nrf54l15/cpuapp/ns as board variant
This commit adds the nRF54l15//nrf54l15/cpuapp/ns board variant to
sdk-zephyr. It allows to build nRF54l15pdk with TF-M in nRF Connect SDK.

This is done as a nrf noup patch as the out-of-tree board extension
feature is not available. When zephyrproject-rtos/zephyr#69548 is
completed, then this commit can be reverted and the variant be placed in
sdk-nrf repository.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>
(cherry picked from commit 8217fd314bd2aab6d1d27ce017e3ab26b79821f3)