[nrf fromlist] scripts: list_boards: Reintroduce Board.dir as @property
Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/72857

In the `Board` class, the `dir` member was renamed to `directories`,
both to indicate that it is now a list (in HWMv2 with board extensions)
and to reflect the addition of the `BOARD_DIRECTORIES` CMake variable.
Considering that the build system also keeps the `BOARD_DIR` variable,
and for the sake of backwards compatibility and brevity, it should be
useful to retain `Board.dir` in Python as well, symmetrically.

Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>
(cherry picked from commit 25fd4b16e8c488792d0d8450e1b822292f94d3c3)